refactor(test): relocate weather mocks to test_helpers/mocks
- Moved all mock files from test/features/weather/mocks/ to test/test_helpers/mocks/ to centralize and standardize test utilities. This includes both source and generated mock files for components such as CitySearchViewModel, FirebaseFunctions, FirestoreWeatherRepository, and WeatherApiClient.
