+++ test/test_helpers/mocks/mock_city_search_view_model.mocks.dart
@@ -1,5 +1,5 @@
 // Mocks generated by Mockito 5.4.4 from annotations
-// in weather_app/test/features/weather/mocks/mock_city_search_view_model.dart.
+// in weather_app/test/test_helpers/mocks/mock_city_search_view_model.dart.
 // Do not manually edit this file.
 
 // ignore_for_file: no_leading_underscores_for_library_prefixes
