+++ test/test_helpers/mocks/mock_firebase_functions.dart
@@ -1,5 +1,5 @@
-import 'package:mockito/annotations.dart';
 import 'package:cloud_functions/cloud_functions.dart';
+import 'package:mockito/annotations.dart';
 
 @GenerateMocks([FirebaseFunctions, HttpsCallable, HttpsCallableResult])
 void main() {}
