+++ test/test_helpers/mocks/mock_weather_api_client.mocks.dart
@@ -1,5 +1,5 @@
 // Mocks generated by Mockito 5.4.4 from annotations
-// in weather_app/test/features/weather/mocks/mock_firestore_weather_repository.dart.
+// in weather_app/test/test_helpers/mocks/mock_weather_api_client.dart.
 // Do not manually edit this file.
 
 // ignore_for_file: no_leading_underscores_for_library_prefixes
@@ -7,7 +7,7 @@ import 'dart:async' as _i4;
 
 import 'package:mockito/mockito.dart' as _i1;
 import 'package:weather_app/core/network/response/result.dart' as _i2;
-import 'package:weather_app/data/repositories/weather/firestore_weather_repository.dart'
+import 'package:weather_app/data/services/weather/weather_api_client.dart'
     as _i3;
 import 'package:weather_app/domain/models/weather/forecast.dart' as _i5;
 
@@ -34,27 +34,26 @@ class _FakeResult_0<T> extends _i1.SmartFake implements _i2.Result<T> {
         );
 }
 
-/// A class which mocks [FirestoreWeatherRepository].
+/// A class which mocks [WeatherApiClient].
 ///
 /// See the documentation for Mockito's code generation for more information.
-class MockFirestoreWeatherRepository extends _i1.Mock
-    implements _i3.FirestoreWeatherRepository {
-  MockFirestoreWeatherRepository() {
+class MockWeatherApiClient extends _i1.Mock implements _i3.WeatherApiClient {
+  MockWeatherApiClient() {
     _i1.throwOnMissingStub(this);
   }
 
   @override
-  _i4.Future<_i2.Result<_i5.Forecast>> fetchForecast(String? cityName) =>
+  _i4.Future<_i2.Result<_i5.Forecast>> fetchWeather(String? cityName) =>
       (super.noSuchMethod(
         Invocation.method(
-          #fetchForecast,
+          #fetchWeather,
           [cityName],
         ),
         returnValue: _i4.Future<_i2.Result<_i5.Forecast>>.value(
             _FakeResult_0<_i5.Forecast>(
           this,
           Invocation.method(
-            #fetchForecast,
+            #fetchWeather,
             [cityName],
           ),
         )),
